requirements.txt
PyYAML
jsonschema
pytest
sphinx
sphinx-rtd-theme
sphinxcontrib-napoleon
